requirements.txt
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==1.9
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.7.0
django-pure-pagination==0.3.0
django-reversion==2.0.13
django-simple-captcha==0.4.6
django-widgets==0.1.16
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.10.3
idna==2.6
jdcal==1.3
mysqlclient==1.3.10
odfpy==1.3.6
openpyxl==2.4.10
Pillow==5.0.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
